Try out shader defs

+++ bevy/snooker/src/config.rs
@@ -4,7 +4,7 @@ pub struct Config {
     pub size: (u32, u32),
     pub workgroup_size: u32,
     pub ball_radius: f32,
-    pub number_of_balls: usize,
+    pub number_of_balls: i32,
     pub wall_width: f32,
     pub wall_color: Color,
     pub pocket_radius: f32
