+++ bevy/snooker/src/plugin.rs
@@ -51,7 +51,7 @@ impl Plugin for GpuComputePlugin {
 
         let ball_buffer = render_device.create_buffer(&BufferDescriptor {
             label: None,
-            size: (std::mem::size_of::<Vec4>() * CONFIG.number_of_balls) as u64,
+            size: (std::mem::size_of::<Vec4>() * (CONFIG.number_of_balls as usize)) as u64,
             usage: BufferUsages::UNIFORM | BufferUsages::COPY_DST,
             mapped_at_creation: false,
         });
